+++ NEWS.md
@@ -1,3 +1,10 @@
+# spant 1.12.0
+* Added 2HG and citrate simulation parameters. e.g. get_mol_paras("2hg").
+* Better print output for molecular definitions.
+* Added metabolite and basis simulation vignettes.
+* Bug fix for setting the ppm reference when reading LCModel RAW files.
+* Bug fix for ABfit CRLB calculation of combined signals, eg tNAA, tCr.
+
 # spant 1.11.0
 * Options added to allow extra information to attached to mrs_data and
 fit_result objects as a data frame.